requirements.txt
requests
scikit-learn == 1.3.0
pandas == 2.0.3
google-cloud-vision==3.4.4
st-annotated-text==4.0.1